+++ lib/components/plan_card.dart
@@ -1,4 +1,4 @@
-import 'package:alpha_fe/pages/plan_page/plan_page_2.dart';
+import 'package:alpha_fe/pages/plan_page/plan_page_2/plan_page_2.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
