+++ lib/components/plan_course_event.dart
@@ -4,8 +4,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:dio/dio.dart';
 import 'package:alpha_fe/components/plan_edit.dart';
-import 'package:alpha_fe/pages/plan_page/plan_page.dart';
-import 'package:alpha_fe/pages/plan_page/plan_page_2.dart';
+import 'package:alpha_fe/pages/plan_page/plan_page_1/plan_page.dart';
+import 'package:alpha_fe/pages/plan_page/plan_page_2/plan_page_2.dart';
 
 import 'date_dropdown.dart';
 //여행 코스
